[ALPS05066801] usb20: fix out-of-bound access in musb_host_active_dev_add
musb_host_active_dev_add

usb20: fix out-of-bound access in musb_host_active_dev_add()
1. replace int to unsigned int.
2. limits the addr number small than table size.

MTK-Commit-Id: d782cac755a9ebeaeb34f517a214ee7e82ff2fae

CR-Id: ALPS05066801
Feature: [Module]USB 2.0

+++ drivers/misc/mediatek/usb20/musb_host.c
@@ -2879,7 +2879,7 @@ static int
 
 #ifdef CONFIG_MTK_MUSB_QMU_SUPPORT
 	if (urb->dev->devnum)
-		musb_host_active_dev_add(urb->dev->devnum);
+		musb_host_active_dev_add((unsigned int)urb->dev->devnum);
 #endif
